+++ integrations/github/github/core/options.py
@@ -1,4 +1,4 @@
-from typing import TypedDict
+from typing import Required, TypedDict
 
 
 class SingleRepositoryOptions(TypedDict):
@@ -11,3 +11,11 @@ class ListRepositoryOptions(TypedDict):
     """Options for listing repositories."""
 
     type: str
+
+
+class SingleUserOptions(TypedDict):
+    login: Required[str]
+
+
+class SingleTeamOptions(TypedDict):
+    slug: Required[str]