+++ integrations/github/CHANGELOG.md
@@ -33,10 +33,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.1.3-dev (2025-06-01)
 
-
 ### Improvements
 
 - Bumped ocean version to ^0.23.5
+
+### Features
+
+- Implement support for User kinds
+- Implement support for Team kinds
 
 
 ## 0.1.2-dev (2025-05-30)